+++ CHANGELOG.md
@@ -11,6 +11,10 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 ### Fixed
 - [cpp] Actually use the VERSION file ([#376](https://github.com/cucumber/gherkin/pull/376))
 - [Java] Prefer the longest step keyword ([#401](https://github.com/cucumber/gherkin/pull/401))
+- [.NET] Prefer the longest step keyword ([#405](https://github.com/cucumber/gherkin/pull/405))
+- [PHP] Prefer the longest step keyword ([#403](https://github.com/cucumber/gherkin/pull/403))
+- [Go] Prefer the longest step keyword ([#403](https://github.com/cucumber/gherkin/pull/404))
+- [JavaScript] Prefer the longest step keyword ([#407](https://github.com/cucumber/gherkin/pull/407))
 - [Ruby] Prefer the longest step keyword ([#409](https://github.com/cucumber/gherkin/pull/409))
 
 ## [32.1.1] - 2025-04-11